// demo/client/src/app/components/ListConfigureState.tsx
import { IIdItem } from "../../models/api/IIdItem";

export interface ListConfigureState<T extends IIdItem> {
    /**
     * The items in the list.
     */
    items: T[];

    /**
     * Is an item being configured.
     */
    configuring: boolean;

    /**
     * The item to delete.
     */
    deleteItem?: T;
}